DragoonCharacter
DragoonCharacter now can toggle off the orient direction to movement to
only be active while the sword is drawn.
DragoonCharacter now will face the same direction as the camera when the
sword is drawn.
TODO change idle/move blendspace from 2D to 1D.
TODO figure out delay in why SwordIdle/Move blendspace has delay in
updating the orientation of the player in blueprints.