Animations and DragoonCharacter_BP
DragoonCharacter_BP now spawns a sword on begin play and attaches it to
the Sword_Sheathe socket on the hip.
Added an event notify, SwordIsDrawn, to the DragoonCharacter_AnimBP
animation of sword equip.
DragoonCharacter_AnimBP will attempt to cast to DragoonCharacter_BP, and
based on the value of bIsSwordDrawn, will set the sword to be attached
to either the hand or hip socket.
Sword now behaves correctly when in either socket.