Animations
Setup animation states for Parry and Dodge.
Added animations for parrying to Parry state.
Parry state uses DirectionOfAttack to determine which parry to play.
Dodge state uses the x and y movement of the character to determine
which dodge should be played.
Added bools to DragoonCharacter for isDodging and isParrying.