Animations and DragoonCharacter
Created animation notify events for FinishedParry and FinishedDodge.
Wired BP for new notify events.
Change dodge state transition rules as SpeedX was incorrectly set for
lateral movement instead of forward/backward and SpeedY was opposite.
(SpeedX is MoveForward and SpeedY is MoveRight)
Added the notify event points in the animations used in parry and dodge
animations.
Removed bDodgeHeld from DragoonCharacter as it is not necessary.
Added ignore move input to parry and dodge functions. Move input is
reset in the finished function for those actions.