AttackCircle and DragoonCharacter_BP
Updated DragoonCharacter_BP to use the MyTakeDamage function instead of
TakeDamage.
Added functions to AttackCircle for an agent joining and leaving the
circle.
Added function to AttackCircle for getting the location of a slot
assigned to an agent.
Added TMap to track the offset of the locations around the center of the
circle.
Added error handling code for when an agent requests an operation but is
not in the circle.
Added code for updating the location of slots based on the circle's
center.
Added function to find the nearest empty spot on circle to a requesting
agent.
Removed circle's rotation as it is irrelevant to the purpose of the
grid.
Added comments to code to clarify what was happening in code and why it
was being done.,
Recompiled all blueprints to make sure everything was running.