PlayerCharacter, EnemyAgent, and AttackCircle
Created new class PlayerCharacter that inherits from DragoonCharacter
and will hold any data specific to the main player in the game.
PlayerCharacter sets up an instance of AttackCircle for itself during
its construction.
Added a function to get the enemy score from an enemy agant to
EnemyAgent.
Added Initialize function with checking for a valid player to attack
circle. This uses what code was previously set in the constructor.
Initialize the occupant TMap for attack circle now in Initialize.
Added function to set player pointer for attack circle.
Added logic to JoinCircle function in attack circle to check whether an
enemy agent can join the circle and change available score if they join.