Attack circle use in multiple classes
Attack circle is now being created in the DragoonGameMode class.
The attack circle receives the player from PlayerCharacter's Begin Play
function. This function also initilaizes the attack circle with initial
data of the player's location.
DragoonAIController now gets a pointer to the character it is
controlling.
PlayerCharacter now stores a reference pointer to the attack circle
object owned by the game mode.