AI class work
Removed the AI Blackboard classes that inherited from parts of UE4's
classes.
Added in the class DragoonAIBlackboard to use as my own version of a
blackboard for the game.
DragoonAIBlackboard has TArrays for enemy agents to keep them as an in
combat and out of combat group. Methods for joining, leaving, and
swapping groups were also made.
DragoonAIBlackboard will eventually own the attack circle instead of the
game mode, but the game mode will own the blackboard class instance.
Added a bool to enemy agents for whether they were in combat or not.
This should be set upon them entering the attack circle or leaving it.
Possibly add another bool for an is alerted state to easily check if an
agent is on the outskirts of a fight when enough enemy score would be
available to insert a new enemy.