Blackboard and Attack struct
Added a struct in DragoonCharacter for recording an attack's direction,
type, and id (id is just a composite of values for direction and type).
Set EnemyAgent to be able to tick. Will be used to override Tick
function for use with own FSM.
Added 3D int array to blackboard to host N grams for player attacks.
Added variables to store last 3 attacks to blackboard.
Added function to record the attack from a player and increment the
total of that N gram in the array.