AI
Made it so AttackCircle is now an object held by Blackboard. Using
shared pointers was resulting in the object being deleted as soon as the
function in which it was constructed was completed.
AIController will have enemy draw sword without crashing. Still unable
to move enemy agents using the MoveToActor or MoveToLocation functions.
NavMesh has been rebuilt multiple times. No longer seem to be crashing,
but movement is not occurring either.