AI, Combat, DragoonCharacter
Changed on overlap event function to be called now inside the .cpp file
for DragoonCharacter. This still performs same checks as the blueprint
did originally.
Made MyTakeDamage virtual so I could overload it in EnemyAgent.
EnemyAgent now sends the information of its death to the blackboard
which relays the information to the AI controller. (cannot send directly
due to circular inheritance problems)
If an enemy agent dies in the circle, circle enemy score is updated and
new enemies will join the circle.
Added recovered from hit notifies to the animBP for SwordHit animations.