AI FSM States
Added a waypoint TArray of FVectors to EnemyAgent if agent would be used
to patrol a certain path.
Added a guard post vector to EnemyAgent if it would be used to guard a
certain location.
Set logic for GuardState to wander around the guard post of an agent
wihtin a certain bounds. Agent will wait for a small amount of time
before going to check out another spot nearby.
Set logic for PatrolState to follow a series of waypoints in a circular
fashion. Agents can be made to wait for a small amount of time before
pursuing the next waypoint.