N Gram
Started working on function for using N grams for predictions.
Currently will get a map with number of occurences from history and one
from the ngram totals.
Need to find a way to get probability from these and determine what the
next attack is most likely to be or a range of likely attacks to select
from based on their weights.